+++ src/protobuf/assembler.test.ts
@@ -1,7 +1,11 @@
-import { lrc } from "./assembler";
+import { gtsm, lrc } from "./assembler";
 
 it("can generate lrc", () => {
   expect(lrc({ videoId: "foo", channelId: "bar" }, { top: true })).toBe(
     "0ofMyAMxGihDZ3dxQ2dvRFltRnlFZ05tYjI4YUMrcW8zYmtCQlFvRFptOXZJQUU9MAGCAQIIBA%3D%3D"
   );
+});
+
+it("can generate gtsm", () => {
+  expect(gtsm()).toBe("CgNhc3ISAmVuGgA%3D");
 });