+++ src/protobuf/assembler.ts
@@ -178,17 +178,21 @@ export function cmp(
   );
 }
 
-export function gts(videoId: string) {
+export function gts(videoId: string, language: string) {
   return b64e(
     cc([
       ld(1, videoId),
-      ld(2, "CgNhc3ISAmlkGgA%3D"),
+      ld(2, gtsm(language)),
       vt(3, 1),
       ld(5, "engagement-panel-searchable-transcript-search-panel"),
       vt(6, 1),
     ]),
     B64Type.B1
   );
+}
+
+export function gtsm(language: string) {
+  return b64e(cc([ld(1, "asr"), ld(2, language), ld(3, "")]), B64Type.B1);
 }
 
 /**